+++ plugins/txtout/txtout.c
@@ -164,7 +164,7 @@ txtout_output(const char *descr, iaddr from, iaddr to, uint8_t proto, unsigned f
 	/*
 	 * IP Stuff
 	 */
-	fprintf(out, "%10ld.%06ld", ts.tv_sec, (long)ts.tv_usec);
+	fprintf(out, "%10ld.%06ld", (long)ts.tv_sec, (long)ts.tv_usec);
 	fprintf(out, " %s %u", ia_str(from), sport);
 	fprintf(out, " %s %u", ia_str(to), dport);
 	fprintf(out, " %hhu", proto);
